refatoração do código gerenciador-de-compras

+++ java/avancado/gerenciador-de-compras/src/br/com/mercado/model/Produto.java
@@ -1,10 +1,10 @@
 package br.com.mercado.model;
 
 public class Produto {
-    private int preco;
+    private double preco;
     private int quantidade;
 
-    public void setPreco(int preco) {
+    public void setPreco(double preco) {
         this.preco = preco;
     }
 
@@ -16,7 +16,7 @@ public class Produto {
         return quantidade;
     }
 
-    public int getPreco() {
+    public double getPreco() {
         return preco;
     }
 }
